requirements.txt
markdown-it-py[plugins]
playwright
requests
abllib
pylint
pytest
